Add fitness scaling operation

+++ src/main/java/pl/edu/pw/ee/evo/operators/impl/StandardEvaluator.java
@@ -17,7 +17,7 @@ public class StandardEvaluator implements Evaluator {
                     .sum());
         }
         var minFitness = population.stream()
-                .mapToInt(Specimen::getFitness)
+                .mapToDouble(Specimen::getFitness)
                 .min().orElseThrow();
         population.forEach(specimen -> specimen.setFitness(specimen.getFitness() - minFitness));
     }
